+++ cloudwanderer/__init__.py
@@ -1,15 +1,15 @@
 """I wandered lonely through the cloud."""
-from . import storage_connectors
-from . import cloud_wanderer_resource
+from . import cloud_wanderer_resource, storage_connectors
+from .aws_interface import CloudWandererAWSInterface
 from .cloud_wanderer import CloudWanderer
-from .aws_urn import AwsUrn
 from .custom_resource_definitions import CustomResourceDefinitions
-from .boto3_interface import CloudWandererBoto3Interface
+from .urn import URN
+
 __all__ = [
-    'storage_connectors',
-    'cloud_wanderer_resource',
-    'CloudWanderer',
-    'AwsUrn',
-    'CustomResourceDefinitions',
-    'CloudWandererBoto3Interface'
+    "storage_connectors",
+    "cloud_wanderer_resource",
+    "CloudWanderer",
+    "URN",
+    "CustomResourceDefinitions",
+    "CloudWandererAWSInterface",
 ]