+++ tests/integration/secondary_attributes/test_secondary_attributes.py
@@ -1,17 +1,17 @@
 import unittest
+
 import boto3
 from parameterized import parameterized
-from cloudwanderer.boto3_interface import CloudWandererBoto3Interface
-from ..mocks import add_infra
+
+from cloudwanderer.aws_interface import CloudWandererAWSInterface
+
 from ..helpers import get_default_mocker, get_secondary_attribute_types
+from ..mocks import add_infra
 
 
 def generate_params():
     get_default_mocker().start_general_mock(restrict_collections=False)
-    services = [
-        ('ec2', 'eu-west-2'),
-        ('iam', 'us-east-1')
-    ]
+    services = [("ec2", "eu-west-2"), ("iam", "us-east-1")]
     for service_name, region_name in services:
         for resource_name, attribute_name in get_secondary_attribute_types(service_name):
             yield (
@@ -19,13 +19,12 @@ def generate_params():
                 service_name,
                 region_name,
                 resource_name,
-                attribute_name
+                attribute_name,
             )
     get_default_mocker().stop_general_mock()
 
 
 class TestSecondaryAttributes(unittest.TestCase):
-
     @classmethod
     def setUpClass(cls):
         get_default_mocker().start_general_mock()
@@ -38,12 +37,9 @@ class TestSecondaryAttributes(unittest.TestCase):
     @parameterized.expand(generate_params())
     def test_query_secondary_attributes(self, _, service_name, region_name, resource_type, attribute_name):
         results = []
-        boto3_interface = CloudWandererBoto3Interface(boto3_session=boto3.Session(region_name=region_name))
+        aws_interface = CloudWandererAWSInterface(boto3_session=boto3.Session(region_name=region_name))
 
-        resources = boto3_interface.get_resources(
-            service_names=[service_name],
-            resource_types=[resource_type]
-        )
+        resources = aws_interface.get_resources(service_names=[service_name], resource_types=[resource_type])
         for resource in resources:
             for secondary_attribute in resource.cloudwanderer_metadata.secondary_attributes:
                 if attribute_name == secondary_attribute.name: